Split exception specs in multiple files
